Add authorization based functionalities. Completed. 8th commit.

+++ 5.8.1/angular/src/app/users/users.component.ts
@@ -1,3 +1,4 @@
+import { AbpSessionService } from 'abp-ng2-module';
 import { Component, Injector } from '@angular/core';
 import { finalize } from 'rxjs/operators';
 import { BsModalService, BsModalRef } from 'ngx-bootstrap/modal';
@@ -16,6 +17,7 @@ import {
 import { CreateUserDialogComponent } from './create-user/create-user-dialog.component';
 import { EditUserDialogComponent } from './edit-user/edit-user-dialog.component';
 import { ResetPasswordDialogComponent } from './reset-password/reset-password.component';
+import { Session } from 'inspector';
 
 class PagedUsersRequestDto extends PagedRequestDto {
   keyword: string;
@@ -29,6 +31,11 @@ class PagedUsersRequestDto extends PagedRequestDto {
 export class UsersComponent extends PagedListingComponentBase<UserDto> {
   users: UserDto[] = [];
   departmentlist: DepartmentDto[] = [];
+  departmentIdAllowed:number = null;
+  rolesAllowedList:string[] = [];
+  allowedAdmin:boolean = false;
+  allowedHr:boolean = false;
+  allowedEmployee:boolean = false;
   keyword = '';
   isActive: boolean | null;
   advancedFiltersVisible = false;
@@ -37,21 +44,39 @@ export class UsersComponent extends PagedListingComponentBase<UserDto> {
     injector: Injector,
     private _userService: UserServiceProxy,
     private _departmentService: DepartmentsServiceProxy,
+    private _abpSessionService: AbpSessionService,
     private _modalService: BsModalService
   ) {
     super(injector);
   }
 
   createUser(): void {
-    this.showCreateOrEditUserDialog();
+    if(this.allowedAdmin || this.allowedHr){
+      this.showCreateOrEditUserDialog();
+    }else{
+      abp.message.info("Sorry! You are not allowed.")
+    }
+  
   }
 
   editUser(user: UserDto): void {
-    this.showCreateOrEditUserDialog(user.id);
+    if(this.allowedAdmin || this.allowedHr){
+      this.showCreateOrEditUserDialog(user.id);
+    }else if(this._abpSessionService.userId === user.id){
+      this.showCreateOrEditUserDialog(user.id);
+    }else{
+      abp.message.info("Sorry! You are not allowed.")
+    }
+   
   }
 
   public resetPassword(user: UserDto): void {
-    this.showResetPasswordUserDialog(user.id);
+    if(this.allowedAdmin || this.allowedHr){
+      this.showResetPasswordUserDialog(user.id);
+    }else{
+      abp.message.info("Sorry! You are not allowed.")
+    }
+    
   }
 
   clearFilters(): void {
@@ -69,8 +94,35 @@ export class UsersComponent extends PagedListingComponentBase<UserDto> {
     request.isActive = this.isActive;
     this._departmentService.listAll().subscribe((result)=>{
       this.departmentlist = result;
-    })
+    });
 
+    this._userService.get(this._abpSessionService.userId).subscribe((result)=>{
+      this.rolesAllowedList = result.roleNames;
+      this.departmentIdAllowed = result.departmentId;
+      result.roleNames.forEach(element => {
+        // console.log(element);
+        
+        if(element.toLocaleLowerCase() == "admin"){
+          this.allowedAdmin = true;
+        }
+        if(element.toLocaleLowerCase() == "hr"){
+          this.allowedHr = true;
+        }
+        if(element.toLocaleLowerCase() == "employee"){
+          this.allowedEmployee = true;
+        }
+      });
+
+      // console.log(this.allowedAdmin);
+      // console.log(this.allowedHr);
+      // console.log(this.allowedEmployee);
+    });
+
+    // console.log(this._abpSessionService.userId);
+    // console.log(typeof(this._abpSessionService.userId));
+    
+    
+    
     this._userService
       .getAll(
         request.keyword,
@@ -85,23 +137,29 @@ export class UsersComponent extends PagedListingComponentBase<UserDto> {
       )
       .subscribe((result: UserDtoPagedResultDto) => {
         this.users = result.items;
+       
         this.showPaging(result, pageNumber);
       });
   }
 
   protected delete(user: UserDto): void {
-    abp.message.confirm(
-      this.l('UserDeleteWarningMessage', user.fullName),
-      undefined,
-      (result: boolean) => {
-        if (result) {
-          this._userService.delete(user.id).subscribe(() => {
-            abp.notify.success(this.l('SuccessfullyDeleted'));
-            this.refresh();
-          });
+    if(this.allowedAdmin || this.allowedHr){
+      abp.message.confirm(
+        this.l('UserDeleteWarningMessage', user.fullName),
+        undefined,
+        (result: boolean) => {
+          if (result) {
+            this._userService.delete(user.id).subscribe(() => {
+              abp.notify.success(this.l('SuccessfullyDeleted'));
+              this.refresh();
+            });
+          }
         }
-      }
-    );
+      );
+    }else{
+      abp.message.info("Sorry! You are not allowed.")
+    }
+    
   }
 
   private showResetPasswordUserDialog(id?: number): void {
